requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.7
Django==5.0.6
django-environ==0.11.2
django-ninja==1.1.0
django-ninja-simple-jwt==0.5.0
gunicorn==22.0.0
packaging==24.0
pip==24.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
setuptools==69.5.1
sqlparse==0.5.0
typing_extensions==4.11.0
wheel==0.43.0
